pyproject.toml
[project]
name = "finance-advisor"
version = "0.1.0"
description = "Personal AI-powered finance advisor with LangGraph + FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.4",
    "langgraph>=1.0.8",
    "aiosqlite>=0.20.0",
    "pydantic-settings>=2.4.0",
    "sse-starlette>=2.1.0",
    "structlog>=24.4.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "pypdf>=4.0.0",
    "yfinance>=0.2.40",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
